requirements.txt
streamlit
leafmap
numpy
matplotlib
pillow
folium
